Move models tests to ocean_lib.
