+++ READMEs/developers.md
@@ -84,7 +84,7 @@ export TEST_PRIVATE_KEY2=0xef4b441145c1d0f3b4bc6d61d29f5c6e502359481152f869247c7
 In console:
 ```console
 #run a single test
-pytest tests/models/test_btoken.py
+pytest ocean_lib/models/tests/test_btoken.py
 
 #run all tests
 pytest
